moving rest of the renderer unit tests over to pytest.
